Split up doc_preprocessors
Closes #63.
