+++ fonduer/parser/__init__.py
@@ -1,27 +1,13 @@
-from fonduer.parser.doc_preprocessors import (
-    CSVPathsPreprocessor,
-    DocPreprocessor,
-    HTMLPreprocessor,
-    TextDocPreprocessor,
-    TSVDocPreprocessor,
-    XMLMultiDocPreprocessor,
-)
 from fonduer.parser.parser import OmniParser
 from fonduer.parser.rule_parser import RegexTokenizer, SpacyTokenizer, Tokenizer
 from fonduer.parser.spacy_parser import Spacy
 
 __all__ = [
-    "CSVPathsPreprocessor",
-    "DocPreprocessor",
-    "HTMLPreprocessor",
     "OmniParser",
     "Parser",
     "ParserConnection",
     "RegexTokenizer",
     "Spacy",
     "SpacyTokenizer",
-    "TSVDocPreprocessor",
-    "TextDocPreprocessor",
     "Tokenizer",
-    "XMLMultiDocPreprocessor",
 ]
