+++ fonduer/parser/preprocessors/__init__.py
@@ -1,0 +1,17 @@
+from fonduer.parser.preprocessors.csv_paths_preprocessor import CSVPathsPreprocessor
+from fonduer.parser.preprocessors.doc_preprocessor import DocPreprocessor
+from fonduer.parser.preprocessors.html_doc_preprocessor import HTMLDocPreprocessor
+from fonduer.parser.preprocessors.text_doc_preprocessor import TextDocPreprocessor
+from fonduer.parser.preprocessors.tsv_doc_preprocessor import TSVDocPreprocessor
+from fonduer.parser.preprocessors.xml_multidoc_preprocessor import (
+    XMLMultiDocPreprocessor
+)
+
+__all__ = [
+    "CSVPathsPreprocessor",
+    "DocPreprocessor",
+    "HTMLDocPreprocessor",
+    "TSVDocPreprocessor",
+    "TextDocPreprocessor",
+    "XMLMultiDocPreprocessor",
+]
