+++ fonduer/parser/preprocessors/doc_preprocessor.py
@@ -1,0 +1,58 @@
+import glob
+import os
+
+
+class DocPreprocessor(object):
+    """
+    Processes a file or directory of files into a set of Document objects.
+
+    :param encoding: file encoding to use, default='utf-8'
+    :param path: filesystem path to file or directory to parse
+    :param max_docs: the maximum number of Documents to produce,
+        default=float('inf')
+
+    """
+
+    def __init__(self, path, encoding="utf-8", max_docs=float("inf")):
+        self.path = path
+        self.encoding = encoding
+        self.max_docs = max_docs
+
+    def generate(self):
+        """
+        Parses a file or directory of files into a set of Document objects.
+
+        """
+        doc_count = 0
+        for fp in self._get_files(self.path):
+            file_name = os.path.basename(fp)
+            if self._can_read(file_name):
+                for doc, text in self.parse_file(fp, file_name):
+                    yield doc, text
+                    doc_count += 1
+                    if doc_count >= self.max_docs:
+                        return
+
+    def __iter__(self):
+        return self.generate()
+
+    def get_stable_id(self, doc_id):
+        return "%s::document:0:0" % doc_id
+
+    def parse_file(self, fp, file_name):
+        raise NotImplementedError()
+
+    def _can_read(self, fpath):
+        return not fpath.startswith(".")
+
+    def _get_files(self, path):
+        if os.path.isfile(path):
+            fpaths = [path]
+        elif os.path.isdir(path):
+            fpaths = [os.path.join(path, f) for f in os.listdir(path)]
+        else:
+            fpaths = glob.glob(path)
+        if len(fpaths) > 0:
+            return fpaths
+        else:
+            raise IOError("File or directory not found: {}".format(path))
