+++ requirements-dev.txt
@@ -1,4 +1,5 @@
 black
+flake8
 isort
 pre-commit
 pytest
